# video/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /douyin/author/video:
        get:
            tags:
                - VideoService
            description: 通过视频 id 获取作者 id
            operationId: VideoService_GetAIdByVId
            parameters:
                - name: videoId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/video.api.video.v1.GetAIdByVIdResp'
...
    /douyin/comment/count/video:
        get:
            tags:
                - CommentService
            description: 获取视频评论总数
            operationId: CommentService_GetCommentCntByVId
            parameters:
                - name: videoId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/comment.api.comment.v1.GetCommentCntByVIdResp'
...
    /douyin/favorite/count/video:
        get:
            tags:
                - FavoriteService
            description: 获取视频的获赞数
            operationId: FavoriteService_GetFavoriteCntByVId
            parameters:
                - name: id
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/favorite.api.favorite.v1.GetFavoriteCntByVIdResponse'
...
    /douyin/favorite/list:
        get:
            tags:
                - FavoriteService
            description: 获取用户点赞视频列表
            operationId: FavoriteService_FavoriteList
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/favorite.api.favorite.v1.Douyin_favorite_list_response'
        post:
            tags:
                - VideoService
            description: 通过喜欢视频 id 列表获取视频列表
            operationId: VideoService_FavoriteListByVId
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/video.api.video.v1.FavoriteListReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/video.api.video.v1.FavoriteListResp'
    /douyin/feed:
        get:
            tags:
                - VideoService
            description: 获取视频流
            operationId: VideoService_Feed
            parameters:
                - name: latestTime
                  in: query
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/video.api.video.v1.Douyin_feed_response'
...
    /douyin/publish/action:
        post:
            tags:
                - VideoService
            description: 发布视频
            operationId: VideoService_Publish
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/video.api.video.v1.Douyin_publish_action_request'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/video.api.video.v1.Douyin_publish_action_response'
    /douyin/publish/list:
        get:
            tags:
                - VideoService
            description: 获取用户发布的视频列表
            operationId: VideoService_PublishList
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/video.api.video.v1.Douyin_publish_list_response'
    /douyin/publish/vids:
        get:
            tags:
                - VideoService
            description: 通过作者 id 获取作者的发布视频 id 列表
            operationId: VideoService_PublishVidsByAId
            parameters:
                - name: authorId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/video.api.video.v1.PublishVidsByAIdResp'
...
    /douyin/work/count:
        get:
            tags:
                - VideoService
            description: 获取用户作品数量
            operationId: VideoService_WorkCnt
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/video.api.video.v1.WorkCnt_response'
components:
    schemas:
...
        comment.api.comment.v1.GetCommentCntByVIdResp:
            type: object
            properties:
                commentCount:
                    type: string
...
        favorite.api.favorite.v1.Douyin_favorite_list_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                videoList:
                    type: array
                    items:
                        $ref: '#/components/schemas/favorite.api.favorite.v1.Video'
...
        favorite.api.favorite.v1.GetFavoriteCntByVIdResponse:
            type: object
            properties:
                favoriteCount:
                    type: string
...
        favorite.api.favorite.v1.User:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                followCount:
                    type: string
                followerCount:
                    type: string
                isFollow:
                    type: boolean
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
                totalFavorited:
                    type: string
                workCount:
                    type: string
                favoriteCount:
                    type: string
        favorite.api.favorite.v1.Video:
            type: object
            properties:
                id:
                    type: string
                author:
                    $ref: '#/components/schemas/favorite.api.favorite.v1.User'
                playUrl:
                    type: string
                coverUrl:
                    type: string
                favoriteCount:
                    type: string
                commentCount:
                    type: string
                isFavorite:
                    type: boolean
                title:
                    type: string
...
        video.api.video.v1.Douyin_feed_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                videoList:
                    type: array
                    items:
                        $ref: '#/components/schemas/video.api.video.v1.Video'
                nextTime:
                    type: string
        video.api.video.v1.Douyin_publish_action_request:
            type: object
            properties:
                token:
                    type: string
                data:
                    type: string
                    format: bytes
                title:
                    type: string
        video.api.video.v1.Douyin_publish_action_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
        video.api.video.v1.Douyin_publish_list_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                videoList:
                    type: array
                    items:
                        $ref: '#/components/schemas/video.api.video.v1.Video'
        video.api.video.v1.FavoriteListReq:
            type: object
            properties:
                videoIdList:
                    type: array
                    items:
                        type: string
        video.api.video.v1.FavoriteListResp:
            type: object
            properties:
                videoList:
                    type: array
                    items:
                        $ref: '#/components/schemas/video.api.video.v1.Video'
        video.api.video.v1.GetAIdByVIdResp:
            type: object
            properties:
                authorId:
                    type: string
        video.api.video.v1.PublishVidsByAIdResp:
            type: object
            properties:
                videoIdList:
                    type: array
                    items:
                        type: string
        video.api.video.v1.User:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                followCount:
                    type: string
                followerCount:
                    type: string
                isFollow:
                    type: boolean
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
                totalFavorited:
                    type: string
                workCount:
                    type: string
                favoriteCount:
                    type: string
        video.api.video.v1.Video:
            type: object
            properties:
                id:
                    type: string
                author:
                    $ref: '#/components/schemas/video.api.video.v1.User'
                playUrl:
                    type: string
                coverUrl:
                    type: string
                favoriteCount:
                    type: string
                commentCount:
                    type: string
                isFavorite:
                    type: boolean
                title:
                    type: string
        video.api.video.v1.WorkCnt_response:
            type: object
            properties:
                workCount:
                    type: string
...